lab2/stripe_test_task-master/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-jazzmin==2.6.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.5
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
stripe==5.0.0
urllib3==1.26.12
